Refactored midi input code to allow unlimited syex length + partial sysex callbacks on win32. Fixed a few problems with menu bars, Quicktime, AudioUnits. Modernised some old win32 file chooser code. Tweaked some window border rendering.

+++ extras/juce demo/Source/demos/OpenGLDemo.cpp
@@ -99,7 +99,7 @@ public:
               << ", " << pixFormat->accumulationBufferBlueBits
               << ", " << pixFormat->accumulationBufferAlphaBits
               << "), full-scene AA="
-              << pixFormat->fullSceneAntiAliasingNumSamples;
+              << (int) pixFormat->fullSceneAntiAliasingNumSamples;
 
             Logger::outputDebugString (formatDescription);
         }
